fix: fix cryptographic vulnerabilities in id masking
Note: I haven’t verified the vulnerabilities on the live app or tested the changes on the live app, but the idea of the initial fix here is to be backwards-compatible.

The vulnerabilities this is intended to fix are:

- Deriving the deterministic identifier using an unkeyed hash: It’s possible to brute-force small integers (like 32-bit sequential ids) from a hash very quickly.

- Using unauthenticated encryption to protect the id: Anyone who knows the original id corresponding to a masked id (which is any attacker who has the hash – see the “unkeyed hash” part) can compute a valid mask for any other id as Mask[0..32] || (Mask[32..] XOR id XOR id′).

- Using a raw string as the key: cuts down the key space to below 112 bits.

- Using a stream cipher on variable-length data: leaks the number of decimal digits of the id.

The first two combined mean that the existing masking implementation doesn’t provide any security – an attacker can cheaply get the id corresponding to any masked id and get a valid masked id corresponding to any id.

This new (and old – I think it’s the same as my original Rust) implementation uses the raw AES block cipher as a pseudorandom permutation to get the desired properties of both `Hash` and `Mask` (and actually just makes them the same), and gets authentication from the leftover bits in the block by requiring them all to be zero. It also uses binary instead of text encoding to avoid side channels (just as a matter of hygiene – everything else ids are used for won’t be so protected), and should be more efficient on several levels (but again, not actually benchmarked yet).

The general output shape is the same for now for some measure of compatibility, but `MASK_SECRET` values need to be updated, and any existing masks will stop working. If possible, I recommend replacing the `{ Id: X, Mask: X }` with just `X` for efficiency.

This new implementation is also restricted to 32-bit `uint` values, since the size of `uint` isn’t fixed and PostgreSQL’s `integer` is already restricted to *signed* 32-bit integers – but it extends easily and with enough safety margin to 64 bits. Any future larger string ids (like user ids) will need a different approach, but the end result can still be a single opaque, deterministic string (with an SIV mode, for example).